+++ pennylane_lightning/core/src/simulators/lightning_tensor/tensor/base/TensorBase.hpp
@@ -48,7 +48,7 @@ template <class PrecisionT, class Derived> class TensorBase {
         length_ = std::accumulate(extents.begin(), extents.end(),
                                   std::size_t{1}, std::multiplies<>());
     }
-
+    // Construct a tensor with given extents
     explicit TensorBase(const std::vector<std::size_t> &extents)
         : rank_(extents.size()),
           modes_(std::move(std::vector(rank_, std::size_t{0}))),