+++ pennylane_lightning/core/src/simulators/lightning_tensor/tncuda/MPSTNCuda.hpp
@@ -32,6 +32,7 @@
 #include "TNCudaBase.hpp"
 #include "TensorCuda.hpp"
 #include "TensornetBase.hpp"
+#include "Util.hpp"
 #include "cuda_helpers.hpp"
 #include "tncudaError.hpp"
 #include "tncuda_helpers.hpp"
@@ -140,13 +141,12 @@ class MPSTNCuda final : public TNCudaBase<Precision, MPSTNCuda<Precision>> {
     /**
      * @brief Get a vector of pointers to tensor data of each site.
      *
-     * @return std::vector<uint64_t *>
+     * @return std::vector<CFP_t *>
      */
-    [[nodiscard]] auto getTensorsOutDataPtr() -> std::vector<int64_t *> {
-        std::vector<int64_t *> tensorsOutDataPtr(BaseType::getNumQubits());
+    [[nodiscard]] auto getTensorsOutDataPtr() -> std::vector<CFP_t *> {
+        std::vector<CFP_t *> tensorsOutDataPtr(BaseType::getNumQubits());
         for (std::size_t i = 0; i < BaseType::getNumQubits(); i++) {
-            tensorsOutDataPtr[i] = reinterpret_cast<int64_t *>(
-                tensors_out_[i].getDataBuffer().getData());
+            tensorsOutDataPtr[i] = tensors_out_[i].getDataBuffer().getData();
         }
         return tensorsOutDataPtr;
     }
@@ -178,8 +178,7 @@ class MPSTNCuda final : public TNCudaBase<Precision, MPSTNCuda<Precision>> {
                         "Please ensure all elements of a basis state should be "
                         "either 0 or 1.");
 
-        CFP_t value_cu =
-            Pennylane::LightningGPU::Util::complexToCu<ComplexT>({1.0, 0.0});
+        CFP_t value_cu = cuUtil::complexToCu<ComplexT>(ComplexT{1.0, 0.0});
 
         for (std::size_t i = 0; i < BaseType::getNumQubits(); i++) {
             tensors_[i].getDataBuffer().zeroInit();
@@ -204,6 +203,9 @@ class MPSTNCuda final : public TNCudaBase<Precision, MPSTNCuda<Precision>> {
         }
     };
 
+    /**
+     * @brief Get final state of the quantum circuit.
+     */
     void get_final_state() {
         if (MPSFinalized_ == MPSStatus::MPSFinalizedNotSet) {
             MPSFinalized_ = MPSStatus::MPSFinalizedSet;
@@ -229,7 +231,7 @@ class MPSTNCuda final : public TNCudaBase<Precision, MPSTNCuda<Precision>> {
             /* const void * */ &algo,
             /* size_t */ sizeof(algo)));
 
-        this->computeState(
+        BaseType::computeState(
             const_cast<int64_t **>(getSitesExtentsPtr().data()),
             reinterpret_cast<void **>(getTensorsOutDataPtr().data()));
     }
@@ -256,7 +258,7 @@ class MPSTNCuda final : public TNCudaBase<Precision, MPSTNCuda<Precision>> {
         void *output_tensorPtr[] = {
             static_cast<void *>(output_tensor.getDataBuffer().getData())};
 
-        this->computeState(nullptr, output_tensorPtr);
+        BaseType::computeState(nullptr, output_tensorPtr);
 
         std::vector<ComplexT> results(output_extent.front());
         output_tensor.CopyGpuDataToHost(results.data(), results.size());
@@ -329,16 +331,10 @@ class MPSTNCuda final : public TNCudaBase<Precision, MPSTNCuda<Precision>> {
     std::vector<std::vector<int64_t>> setSitesExtents_int64_() {
         std::vector<std::vector<int64_t>> localSitesExtents_int64;
 
-        for (std::size_t i = 0; i < BaseType::getNumQubits(); i++) {
-            // Convert datatype of sitesExtents to int64 as required by
-            // cutensornet backend
-            std::vector<int64_t> siteExtents_int64(sitesExtents_[i].size());
-            std::transform(sitesExtents_[i].begin(), sitesExtents_[i].end(),
-                           siteExtents_int64.begin(), [](std::size_t x) {
-                               return static_cast<int64_t>(x);
-                           });
-
-            localSitesExtents_int64.push_back(std::move(siteExtents_int64));
+        for (const auto &siteExtents : sitesExtents_) {
+            localSitesExtents_int64.push_back(
+                std::move(Pennylane::Util::cast_vector<std::size_t, int64_t>(
+                    siteExtents)));
         }
         return localSitesExtents_int64;
     }