+++ .github/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release 0.37.0-dev
 
 ### New features since last release
+* Add `observable` and `measurement` support to `cutensornet` backed `lightning.tensor` C++ layer.
+  [(#728)](https://github.com/PennyLaneAI/pennylane-lightning/pull/728)
 
 * Add gate support to `cutensornet` backed `lightning.tensor` C++ layer.
   [(#718)](https://github.com/PennyLaneAI/pennylane-lightning/pull/718)
@@ -9,6 +11,9 @@
   [(#704)](https://github.com/PennyLaneAI/pennylane-lightning/pull/704)
 
 ### Breaking changes
+
+* Removed the `QuimbMPS` class and the corresponding interface/backend from `lightning.tensor`.
+  [(#737)](https://github.com/PennyLaneAI/pennylane-lightning/pull/737)
 
 * Changed the name of `default.tensor` to `lightning.tensor` with the `quimb` backend.
   [(#730)](https://github.com/PennyLaneAI/pennylane-lightning/pull/730)
@@ -49,6 +54,9 @@
 
 ### Bug fixes
 
+* Revert changes calling `IMAG`, `ONE`, `ZERO` templated functions in Kokkos kernels since they are incompatible with device execution.
+  [(#733)](https://github.com/PennyLaneAI/pennylane-lightning/pull/733)
+
 * The `.github/workflows/tests_lkcpu_python.yml` workflow properly checks out the release or stable version of Lightning-Qubit during the test job.
   [(#723)](https://github.com/PennyLaneAI/pennylane-lightning/pull/723)
 
@@ -57,6 +65,9 @@
 
 * Remove the Autoray dependency from requirement files.
   [(#736)](https://github.com/PennyLaneAI/pennylane-lightning/pull/736)
+
+* Fix the `cuda-runtime-12-0` dependency issue on RHEL8.
+  [(#739)](https://github.com/PennyLaneAI/pennylane-lightning/pull/739)
 
 ### Contributors